Fixes #31 - JsMinifier can now safely be registered as a singleton in the container, it creates instances of JsMin on the fly.
